+++ fair_server/src/main/java/com/furb/projeto/dtos/ProdutoLojaDto.java
@@ -1,18 +1,18 @@
 package com.furb.projeto.dtos;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
 
 public class ProdutoLojaDto {
 
     private Integer fkProduto;
     private Integer fkLoja;
 
-    private BigInteger preco;
+    private BigDecimal preco;
 
     public ProdutoLojaDto() {
     }
 
-    public ProdutoLojaDto(Integer fkProduto, Integer fkLoja, BigInteger preco) {
+    public ProdutoLojaDto(Integer fkProduto, Integer fkLoja, BigDecimal preco) {
         this.fkProduto = fkProduto;
         this.fkLoja = fkLoja;
         this.preco = preco;
@@ -34,11 +34,11 @@ public class ProdutoLojaDto {
         this.fkLoja = fkLoja;
     }
 
-    public BigInteger getPreco() {
+    public BigDecimal getPreco() {
         return preco;
     }
 
-    public void setPreco(BigInteger preco) {
+    public void setPreco(BigDecimal preco) {
         this.preco = preco;
     }
 }
